Moved JavaSound into pkgJavaSound
