+++ Slideshow-Creator/src/pkgJavaSound/ControlContext.java
@@ -1,5 +1,6 @@
+package pkgJavaSound;
 /*
- * @(#)JavaSoundApplet.java	1.2	00/01/31
+ * @(#)ControlContext.java	1.5	99/11/01
  *
  * Copyright (c) 1998, 1999 by Sun Microsystems, Inc. All Rights Reserved.
  * 
@@ -28,38 +29,11 @@
  * redistribute the Software for such purposes.
  */
 
-import javax.swing.JApplet;
 
 /**
- * A demo that shows JavaSound features. 
- *
- * Parameters that can be used in the JavaSound.html file inside
- * the applet tag to customize demo runs :
- *            <param name="dir" value="audioDirectory">
- *
- * @(#)JavaSoundApplet.java	1.2	00/01/31
- * @author Brian Lichtenwalter 
+ * The interface for the JavaSound tabs to open and close audio resources.
  */
-public class JavaSoundApplet extends JApplet {
-
-    static JavaSoundApplet applet;
-    private JavaSound demo;
-
-    public void init() {
-        applet = this;
-        String media = "./audio";
-        String param = null;
-        if ((param = getParameter("dir")) != null) {
-            media = param;
-        } 
-        getContentPane().add("Center", demo = new JavaSound(media));
-    }
-
-    public void start() {
-        demo.open();
-    }
-
-    public void stop() {
-        demo.close();
-    }
+public interface ControlContext {
+        public void open();
+        public void close();
 }
