+++ Slideshow-Creator/src/pkgJavaSound/Groove.java
@@ -1,3 +1,4 @@
+package pkgJavaSound;
 /*
  * @(#)Groove.java	1.16	99/11/03
  *
