+++ Slideshow-Creator/src/pkgJavaSound/JavaSoundApplet.java
@@ -1,5 +1,6 @@
+package pkgJavaSound;
 /*
- * @(#)ControlContext.java	1.5	99/11/01
+ * @(#)JavaSoundApplet.java	1.2	00/01/31
  *
  * Copyright (c) 1998, 1999 by Sun Microsystems, Inc. All Rights Reserved.
  * 
@@ -28,11 +29,38 @@
  * redistribute the Software for such purposes.
  */
 
+import javax.swing.JApplet;
 
 /**
- * The interface for the JavaSound tabs to open and close audio resources.
+ * A demo that shows JavaSound features. 
+ *
+ * Parameters that can be used in the JavaSound.html file inside
+ * the applet tag to customize demo runs :
+ *            <param name="dir" value="audioDirectory">
+ *
+ * @(#)JavaSoundApplet.java	1.2	00/01/31
+ * @author Brian Lichtenwalter 
  */
-public interface ControlContext {
-        public void open();
-        public void close();
+public class JavaSoundApplet extends JApplet {
+
+    static JavaSoundApplet applet;
+    private JavaSound demo;
+
+    public void init() {
+        applet = this;
+        String media = "./audio";
+        String param = null;
+        if ((param = getParameter("dir")) != null) {
+            media = param;
+        } 
+        getContentPane().add("Center", demo = new JavaSound(media));
+    }
+
+    public void start() {
+        demo.open();
+    }
+
+    public void stop() {
+        demo.close();
+    }
 }
